fabsf() is not portable.
git-svn-id: https://dev.csoft.net/agar/trunk@5380 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ vg/vg_circle.c
@@ -81,7 +81,7 @@ intsect(VG *vg, VG_Element *vge, double x, double y)
 	vtx = &vge->vtx[0];
 	VG_Car2Pol(vg, x - vtx->x, y - vtx->y, &rho, &theta);
 
-	return (fabsf(rho - vge->vg_circle.radius));
+	return ((float)fabs(rho - vge->vg_circle.radius));
 }
 
 const VG_ElementOps vgCircleOps = {
